pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lmm-classifier"
version = "0.1.0"
description = "Train-free hybrid image classification and retrieval pipeline"
authors = [
    {name = "AutoML Agent"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "scikit-learn",
    "transformers",
    "torch",
    "Pillow",
    "pyyaml",
]

[project.scripts]
build-siglip-index = "lmm_classifier.index_builder:main"

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
